requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
debugpy==1.4.1
decorator==5.0.9
dominate==2.6.0
entrypoints==0.3
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
idna==2.10
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==7.0.2
jupyter-core==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
nest-asyncio==1.5.1
numpy==1.20.2
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
Pygments==2.10.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.2.1
requests==2.25.1
selenium==3.141.0
shiboken2==5.15.2
six==1.15.0
tornado==6.1
traitlets==5.1.0
urllib3==1.26.5
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.0.1
gunicorn==20.0.4
Authlib==0.15.4
python-dotenv==0.19.0
psycopg2-binary==2.9.1